Update Arc - Server Deployment.json
Removed Subscription filter. SubscriptionID was hard coded in the code. 
Added Server Cores to SQL Arc View to have a chance to compare SQL vCore and Server Core. SQL vCore is the logical core and we need to consider Server Core for Paid Cores. HT cores are not counted as Paid cores.